+++ examples/controlpanel/components/event/EventList.jsx
@@ -19,7 +19,7 @@ const EventRow = ({ type, created, _id }) => {
   );
 };
 
-const EventList = ({ onFilterChange, selectTypes, loading, updateHasMore, queryOptions, ...rest }) => {
+const EventList = ({ onFilterChange, types, loading, updateHasMore, queryOptions, ...rest }) => {
   return (
     <InfiniteDataTable
       {...rest}